Add core Algo to split WP as well as its TP
The split of one WorkPeriod will create a new WorkPeriod repr the 1st
half and mutate original WorkPeriod to repr the 2nd half.

The TaskParts within original WorkPeriod requires proper handle as
well. Especially the split Timestamp may in the middle of a TaskPart.

Then again the 1st half of TaskPart goes with newly created WorkPeriod,
the remaining 2nd half will be the 1st TaskPart in the mutated
WorkPeriod, i.e., the original WorkPeriod.