Add the simplest calendar test case
A Calendar is an input container class which end-user should prepare.
While the Schedule class contains the final output to orignized the
Tasks, i.e., TaskParts as well as all WorkPeriods in proper way.

Besides, the TimeZone info of end-user is required so as to generate the
schedule result in a user-friendly time format.

All Tasks should be added through the calendar. The same applies to the
WorkPeriods. It's worth noting that WorkPeriods shall be ordered by each
startTime.